feat: move cartesian animation to its own gallery section with docs references
- Extract cartesian animation from Advanced Patterns into its own
  generate_meta_cartesian_animation.py generator and gallery section
- Add "Cartesian Animation" entry to SECTION_GROUPS in generate_examples.py
- Reference the animation from concepts.md (Design section) and intro.md
  (Adding Dimensions section) so readers can find it from the docs
- Update CLAUDE.md naming convention table with cartesian_ prefix
