+++ AGENTS.md
@@ -118,6 +118,7 @@ are:
 | `publish_` | `publishing/` | example |
 | `rerun_` | `rerun/` | example |
 | `agg_` | `aggregation/` | aggregation form, agg_fn |
+| `cartesian_` | `cartesian_animation/` | (single example) |
 
 **Rules for adding new generators:**
 1. Pick a short, unique section prefix that does not collide with existing prefixes.
